Use PopupButton to get name for query
- Change UI
- Redirect to NewQuery when current query is deleted
- Use history to redirect instead of Redirect component
- Fetch queries even when not applied
- Fix logic for apply and reset disable
- Move save query on view component
- Move filter list query inside "my saved queries" component
- Remove sub-navbar
- Show current query name on header
- Remove "query" edit button with link
- Apply "search" only after "apply" is clicked
