+++ src/components/Navbar/styles.css
@@ -6,16 +6,10 @@
     .bottom {
         display: flex;
         padding: 0 var(--spacing-large);
-
-        .new-entry-link {
-            margin-top: var(--spacing-medium);
-            text-transform: uppercase;
-        }
     }
 
     .top {
         display: flex;
-        border-bottom: var(--width-separator-thin) solid rgba(255, 255, 255, .3);
         padding: 0 calc(var(--spacing-large) - var(--spacing-medium));
 
         .app-brand {
@@ -61,9 +55,19 @@
         }
 
         .actions {
+            display: flex;
             flex-shrink: 0;
             padding: 0 var(--spacing-medium);
-            font-size: var(--font-size-large);
+
+            .new-entry-link {
+                align-self: center;
+                text-transform: uppercase;
+            }
+
+            .dropdown {
+                margin-left: var(--spacing-small);
+                font-size: var(--font-size-large);
+            }
         }
     }
 }
