+++ src/views/Extraction/ExtractionEntriesTable/styles.css
