+++ src/views/Extraction/SavedFiltersList/styles.css
@@ -1,0 +1,30 @@
+.container {
+    display: flex;
+    justify-content: space-between;
+    margin: var(--spacing-medium);
+
+    .search-input {
+        margin-bottom: var(--spacing-medium);
+    }
+
+    .clear-button {
+        margin: calc(-1 * var(--spacing-extra-small));
+        padding: 0;
+        padding: var(--spacing-extra-small);
+
+        .child-container {
+            padding: 0;
+        }
+    }
+
+    .content {
+        display: flex;
+        position: relative;
+        flex-direction: column;
+
+        .selected-query {
+            background: var(--color-background);
+        }
+    }
+}
+
