+++ src/views/Extraction/styles.css
@@ -1,9 +1,50 @@
+.popup {
+    padding: calc(var(--spacing-medium) - var(--spacing-small));
+    width: 400px;
+
+    .comment {
+        margin: var(--spacing-small);
+    }
+
+    .popup-content {
+        display: flex;
+    }
+}
+
 .extraction {
+    display: flex;
     padding: var(--spacing-large);
     width: 100%;
 
     .container {
         margin: var(--spacing-medium);
         min-height: 200px;
+    }
+
+    .large-container {
+        margin: var(--spacing-medium);
+        min-height: 400px;
+    }
+
+
+    .main-content {
+        flex-grow: 1;
+        padding: calc(var(--spacing-medium) - var(--spacing-small));
+        overflow-x: auto;
+    }
+
+    .side-content {
+        display: flex;
+        flex-direction: column;
+        flex-shrink: 0;
+        padding: calc(var(--spacing-medium) - var(--spacing-small));
+        width: calc(160px + 10vw);
+
+        .new-query-button {
+            margin: 0 var(--spacing-medium);
+            border-radius: var(--radius-card);
+            background-color: var(--color-foreground);
+            padding: var(--spacing-medium) var(--spacing-small);
+        }
     }
 }
